+++ test/czech_post_b2b_client_test.rb
@@ -14,16 +14,16 @@ class CzechPostB2bClientTest < Minitest::Test
   def test_knows_available_templates
     all_templates = CzechPostB2bClient::PrintingTemplates.all_classes.to_a
 
-    assert_equal 39, all_templates.size
+    assert_equal 38, all_templates.size
 
     template_61 = all_templates.detect { |t| t.id == 61 }
-    assert_equal 'CP72 - cenný balík do zahraničí (2x A4)', template_61.description
+    assert_equal 'CP72 - cenný balík do zahraničí : 2x', template_61.description
   end
 
   def test_knowns_available_services
     all_services = CzechPostB2bClient::PostServices.all_classes.to_a
 
-    assert_equal 90, all_services.size
+    assert_equal 89, all_services.size
 
     service_32 = all_services.detect { |t| t.code == '32' }
     # assert service_32.# CertificateOfDeliveryWithDeliverToAdresseeOnly
@@ -34,7 +34,7 @@ class CzechPostB2bClientTest < Minitest::Test
   def test_knowns_response_codes
     all_codes = CzechPostB2bClient::ResponseCodes.all_classes.to_a
 
-    assert_equal 263, all_codes.size
+    assert_equal 262, all_codes.size
 
     code_396 = all_codes.detect { |t| t.code == 396 }
     assert_equal 'INFO_CANCEL_SERVICE_5C', code_396.text